+++ Side.php
@@ -313,7 +313,7 @@ $addJavascripts .= 'var menuStudentID="' . UserStudentID() . '",
 		<span class="username br-after"><?php echo User( 'NAME' ); ?></span>
 		<?php
 			// Localized today's date.
-			echo '<span class="today-date size-1">' . strftime( '%A ' . Preferences( 'DATE' ) ) . '</span>';
+			echo '<span class="today-date size-1">' . strftime_compat( '%A ' . Preferences( 'DATE' ) ) . '</span>';
 		?>
 		<br />
 		<?php // School SELECT (Admins & Teachers only).