+++ modules/Accounting/includes/Dashboard.inc.php
@@ -116,7 +116,8 @@ if ( ! function_exists( 'DashboardAccountingAdmin' ) )
 		foreach ( (array) $accounting_RET as $year_month => $accounting )
 		{
 			// Remove dummy day from proper date.
-			$proper_month_year = strftime(
+			// @since 9.0 Fix PHP8.1 deprecated strftime() use strftime_compat() instead
+			$proper_month_year = strftime_compat(
 				trim( str_replace( [ '%d', '//' ], [ '', '/'], Preferences( 'DATE' ) ), '-./ ' ),
 				strtotime( $year_month . '-29' )
 			);